tcpfw: Fix missing return values and subsequent heap corruption
tcpsocks_acceptor() and tcpfw_acceptor() did not return values, and this
occasionally caused tcpfw_listen() to free the tcpfw_t struct while it
was in use.

Make the acceptor functions void, and pass all responsibility for closing
connections and freeing the struct into the acceptor function (or the
threads it creates).

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>