Don't croak if the lwIP TCP queue gets too long
On a slow connection with many pending local->VPN transmissions,
tcp_sndbuf() may indicate that it is safe to write data even though the
queue length is too long.  In this case tcp_write() will return ERR_MEM,
and ocproxy will die.

Add a (somewhat hacky) check for this to local_data_cb().

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>