Import build-debian.sh script from stoken
This automates the process of publishing an ocproxy snapshot to an
Ubuntu PPA repo.  Sample usage:

    make dist
    ./build-debian.sh
    dput ppa:cernekee/ppa tmp.debian/*_source.changes

The script takes care of making sure each upload has a unique "orig"
tarball name, as the Debian system is very sensitive to dissimilar
uploads with the same name.

Also note that in the Debian packaging system, the use of the tilde
ensures that e.g. "ocproxy-0.2~20140614" counts as a LOWER version number
than "ocproxy-0.2".  So when the official "ocproxy-0.2" is released, it
will supersede the old PPA prerelease builds.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>