Re-bind /etc/resolv.conf if the underlying file disappears
NetworkManager regenerates resolv.conf on network changes, which causes
the bind mount inside our namespace to be lost.  On Ubuntu we use
overlayfs to work around the problem, because Ubuntu kernels let
unprivileged users create overlayfs mounts inside their own namespaces.
Standard upstream kernels do not allow this, so they are stuck with bind
mounts.

If bind mounts are used, set up an inotify DELETE_SELF check on
/etc/resolv.conf and have the watcher process reinstitute the bind mount
if it is deleted.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>