vpnns: Create the initial watcher connection before forking
Currently the sequence looks like:

    [parent] fork child
    [child] close fd's / daemonize / etc.
    [parent] attempt connection to child; fail
    [parent] sleep 1 second
    [child] listen on socket
    [parent] wake up from sleep and connect to socket

We can avoid the retry by creating a socketpair prior to forking, and
having the watcher process add its side of the socketpair as a
(special case) watch event.  This eliminates the need to retry.

Also, it means that if the connection to an existing watcher fails,
vpnns can recover somewhat-gracefully by forking a new watcher process.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>